[refact] 즐겨찾기 모달 창, style 분리
